Fix a bug in tau_estimation()
The change is as below
```python
-        a_est = optimize.minimize_scalar(ar1_fun, method='brent').x
+        a_est = optimize.minimize_scalar(ar1_fun, bounds=[0, 1], method='bounded').x
```
Now, the estimation of `a` is bounded within [0, 1] and won't give
inproper result for tau estimation.

Besides:
- The `setup.py` has been updated so that if people install it with
`python setup install`, it will not get an `.egg` file but a folder so
that the `.so` file can be imported correctly.

- In `Spectral.py`, in those functions for users: `wwz()` and
`wwz_psd()`, the NaNs in timeseris (along with the time points) will be
deleted automatically if there is any.

- Add `.gitignore`.