Add holdon as alias of mute in plot funcs
Pyleoclim now assumes %matplotlib inline turned on to be consistent with
other visualization packages, avoiding repeatedly displaying figures in
certain cases.

As a result, any graph object will be displayed automatically when being
created (the behavior of the '%matplotlib inline' mode). Therefore, `mute=True`
will not really prevent displaying the figure, but only hold on
`plt.show()`. In this sense, we should call it `holdon` instead of `mute`.

Therefore, the argument `holdon=None` is added to the plotting function.
`mute` should be deprecated in the future, but is still there to avoid
breaking any existing code. The docstrings code examples have been
updated to use `holdon` though.